Update README.md
SNHU CS-230 Portfolio

Briefly summarize The Gaming Room client and their software requirements. Who was the client? What type of software did they want you to design?

The Gaming Room, a third-party company, hired me to design a document outlining the requirements for porting their game onto mobile devices. The client wanted the game to be available on iPhone and Android platforms.

What did you do particularly well in developing this documentation?

During the development of the documentation, I excelled at providing concise answers to the client's questions. I made sure to be direct and clear when presenting the pros and cons of each platform for their application, making the document easy to read and saving the client's time.

What about the process of working through a design document did you find helpful when developing the code?

Working with a design document kept me on track with the necessary implementation, preventing me from developing unnecessary features that could have wasted time and resources.

If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it?

If I had to revise one part of my work on the documents, I would revisit the choice of language used for the software development. While Java was the most practical choice at the time, after more research, there may be better options available to improve application performance such as Python or C++, especially on mobile devices. Choosing a language that results in a less resource-intensive application would provide a better experience for the game's users.

How did you interpret the user’s needs and implement them into your software design? Why is it so important to consider the user’s needs when designing?

nderstanding the user's needs is crucial when designing software. Failing to consider their requirements can result in wasted time and money, negative reviews, and potential lawsuits. In this case, I interpreted the client's needs and implemented them into the software design to create a product that met their requirements.

How did you approach designing software? What techniques or strategies would you use in the future to analyze and design a similar software application?

When approaching software design, it's essential to fully understand the customer's needs and have open communication with them to ensure all requirements are met. Knowing the customer's priorities and their order of importance is crucial in creating the final design document. In this case, I did not have a physical customer to communicate with, which made it difficult to determine the order of priorities, resulting in a design that may have been different if cost was not a significant concern. In the future, I would use techniques such as regular meetings and user testing to better understand the customer's needs and ensure the final product meets their requirements.